# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type LabAnalysisResponseMethodology = {
  train_test_split: string;
  random_state: number;
  periodic_encoding: string;
  importance_method: string;
};

export type LabModelAnalysisId = typeof LabModelAnalysisId[keyof typeof LabModelAnalysisId];


export const LabModelAnalysisId = {
  dolar: 'dolar',
  glucosa: 'glucosa',
  energia: 'energia',
} as const;

export interface LabFeatureImpact {
  feature: string;
  label: string;
  coefficient: number;
  standardized_impact: number;
  importance_pct: number;
  direction: string;
  interpretation: string;
}

export type LabVisualizationPointsItem = {
  x: number;
  y: number;
};

export interface LabVisualization {
  feature: string;
  label: string;
  points: LabVisualizationPointsItem[];
}

export type LabModelAnalysisMetrics = {
  mse: number;
  rmse: number;
  r2: number;
};

export type LabModelAnalysisDataPreviewItem = {[key: string]: number};

export interface LabModelAnalysis {
  id: LabModelAnalysisId;
  name: string;
  target: string;
  unit: string;
  row_count: number;
  test_row_count: number;
  metrics: LabModelAnalysisMetrics;
  intercept: number;
  feature_impacts: LabFeatureImpact[];
  visualizations: LabVisualization[];
  data_preview: LabModelAnalysisDataPreviewItem[];
  conclusion: string;
}

export interface LabAnalysisResponse {
  models: LabModelAnalysis[];
  methodology: LabAnalysisResponseMethodology;
}

export type LabPredictionRequestModelId = typeof LabPredictionRequestModelId[keyof typeof LabPredictionRequestModelId];


export const LabPredictionRequestModelId = {
  dolar: 'dolar',
  glucosa: 'glucosa',
  energia: 'energia',
} as const;

export type LabPredictionRequestFeatures = {[key: string]: number};

export interface LabPredictionRequest {
  model_id: LabPredictionRequestModelId;
  features: LabPredictionRequestFeatures;
}

export interface LabPredictionResponse {
  model_id: string;
  prediction: number;
  target: string;
  unit: string;
}

/**
 * Entrenar con los tres archivos CSV incluidos en el laboratorio
 */
export type LabTrainingInputSource = typeof LabTrainingInputSource[keyof typeof LabTrainingInputSource];


export const LabTrainingInputSource = {
  included_csv: 'included_csv',
} as const;

export interface LabTrainingInput {
  /** Entrenar con los tres archivos CSV incluidos en el laboratorio */
  source: LabTrainingInputSource;
}