# lib/api-zod/src/generated/types/labTrainingInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { LabTrainingInputSource } from './labTrainingInputSource';

export interface LabTrainingInput {
  /** Entrenar con los tres archivos CSV incluidos en el laboratorio */
  source: LabTrainingInputSource;
}